// src/type.d.ts
export enum ModalType {
  Modal,
  DrawerLeft,
  DrawerRight,
  DrawerTop,
  DrawerBottom,
  ModalTopLeft,
  ModalLeftTop,
  ModalRightTop,
  ModalTopRight,
  ModalLeftBottom,
  ModalBottomLeft,
  ModalRightBottom,
  ModalBottomRight
}

export enum Position {
  LeftTop,
  LeftBottom,
  RightTop,
  RightBottom
}

export enum DirectionType {
  Top,
  Bottom,
  Left,
  Right
}